Version 1.2.7 - fix h.264 and mp4v frame rates in mp4creator - added code from Laurent Aimar for h.264 ctts calculations, fixed problem with multiple nal frames in mp4creator

+++ lib/mp4av/mp4av_mpeg4.h
@@ -69,7 +69,7 @@ bool MP4AV_Mpeg4CreateVol(
 	u_int8_t** ppBytes,
 	u_int32_t* pNumBytes,
 	u_int8_t profile,
-	float frameRate,
+	double frameRate,
 	bool shortTime,
 	bool variableRate,
 	u_int16_t width,
