+++ server/mp4creator/mp4creator.h
@@ -56,7 +56,7 @@
 
 MP4CREATOR_GLOBAL char* ProgName;
 MP4CREATOR_GLOBAL u_int32_t Verbosity;
-MP4CREATOR_GLOBAL float VideoFrameRate;
+MP4CREATOR_GLOBAL double VideoFrameRate;
 MP4CREATOR_GLOBAL u_int32_t Mp4TimeScale;
 MP4CREATOR_GLOBAL bool TimeScaleSpecified;
 MP4CREATOR_GLOBAL bool VideoProfileLevelSpecified;
